# change: Changed run tag filtering query for better performance.