--- a/python_modules/dagster/dagster/_core/storage/alembic/versions/043_add_run_tags_run_id_index.py
+++ b/python_modules/dagster/dagster/_core/storage/alembic/versions/043_add_run_tags_run_id_index.py
@@ -28,7 +28,7 @@ def upgrade():
             unique=False,
             postgresql_concurrently=True,
             mysql_length={
-                "pipeline_name": 512,
+                "run_id": 255
             },
         )